requirements.txt
numpy >= 1.18.0; sys_platform == 'darwin' and python_version > "3.0"
numpy == 1.16.6; sys_platform == 'darwin' and python_version < "3.0"
numpy; sys_platform != 'darwin'
pytz
python-dateutil
ruamel.yaml
multiprocess
mmcif >= 0.72
rcsb.utils.validation >= 0.20
backports.tempfile; python_version < "3.0"
PyNaCl >= 1.3.0
requests >= 2.25
paramiko >=2.7.2
psutil >= 5.7.2
GitPython >= 3.1.18
